finishing up offsets and initial debugging, close to merging with dev
