+++ fedcal/time_utils.py
@@ -17,9 +17,27 @@ time conversions in fedcal. We expose it publicly because they're probably
 generally useful for other things.
 
 It includes:
-- `get_today_in_posix_day` and `pdtimestamp_to_posix_day` for
+- `iso_to_ts` streamlines conversion of ISO8601 formatted strings to pandas
+Timestamp, primarily for loading json status input in _status_factory.py
+
+- `get_today` and `ts_to_posix_day` for
 handling POSIX-day integer second retrieval for today and from pandas
 Timestamp objects respectively.
+
+- `ensure_datetimeindex` a decorator that normalizes a variety of input to
+DatetimeIndex for consistent handling
+
+- `to_dt64` a decorator that normalizes a variety of input to numpy.datetime64
+for consistent handling. Primarily used for back-end custom DateOffset object
+calculations.
+
+- `dt64_to_date` converts an array (or scalar) of numpy.datetime64 to integer
+y, m, d using numpy vectorized operations. Primarily used for custom
+DateOffset object calculations.
+
+- `dt64_to_dow` converts an array (or scalar) of numpy.datetime64 to integer
+dow using numpy vectorized operations. Primarily used for custom DateOffset
+object calculations.
 
 - `YearMonthDay` a class for handling date conversions from year, month, day.
 It's there because I wanted something cleaner than datetime. I like it. It's
@@ -53,22 +71,27 @@ input to U.S. Eastern, as with `to_timestamp`.
 from __future__ import annotations
 
 import datetime
-from enum import unique
-from functools import singledispatch, wraps
-from typing import Any, Callable, Self
+from functools import singledispatch
+from typing import Any, Self
 
 import numpy as np
 import pandas as pd
 from attrs import astuple, define, field
+from funcy import decorator
 from numpy import datetime64, int32, int64, uint8
 from numpy.typing import NDArray
 from pandas import DatetimeIndex, Index, PeriodIndex, Series, Timestamp
 from pandas.tseries.frequencies import to_offset
 
-from fedcal._typing import FedIndexConvertibleTypes, FedStampConvertibleTypes
+from fedcal._typing import (
+    DatetimeScalarOrArray,
+    FedIndexConvertibleTypes,
+    FedStampConvertibleTypes,
+    TimestampSeries,
+)
 
 
-def to_dt(t: str, fmt: str | None = None) -> Timestamp:
+def iso_to_ts(t: str, fmt: str | None = None) -> Timestamp:
     """
     Short and quick string to datetime conversion specifically for
     loading intervals.
@@ -77,7 +100,7 @@ def to_dt(t: str, fmt: str | None = None) -> Timestamp:
     ----------
     t : str
         The string to convert.
-    fmt : str, optional, defaults to
+    fmt : str, optional, defaults to ISO08601
     Returns
     -------
     Timestamp
@@ -98,7 +121,7 @@ def get_today() -> Timestamp:
     return pd.Timestamp.utcnow().normalize()
 
 
-def pdtimestamp_to_posix_day(timestamp: Timestamp) -> int:
+def ts_to_posix_day(timestamp: Timestamp) -> int:
     """
     Converts a pandas Timestamp object to a POSIX-day integer timestamp.
 
@@ -115,29 +138,88 @@ def pdtimestamp_to_posix_day(timestamp: Timestamp) -> int:
     return int(timestamp.normalize().timestamp() // 86400)
 
 
-def ensure_datetimeindex(
-    dates: Series[Timestamp] | Timestamp | DatetimeIndex,
-) -> DatetimeIndex:
+@decorator
+def ensure_datetimeindex(call) -> Any:
     """
-    Simple function to ensure a DatetimeIndex for certain calculations.
+    Decorator to ensure that the argument is a DatetimeIndex before calling
+    the original function.
 
     Parameters
     ----------
-    dates
-        Date or dates to ensure are outputted to a DatetimeIndex.
+    call : calling function
 
     Returns
     -------
-        DatetimeIndex
+    Wrapped func output from converted datetime index.
     """
-    if isinstance(dates, (pd.Series, pd.Timestamp)):
-        return pd.DatetimeIndex(
-            data=[dates] if isinstance(dates, pd.Timestamp) else dates
+    method = hasattr(call._args[0], "__class__")
+    dates = call._args[1] if method else call._args[0]
+    other_args = call._args[2:] if method else call._args[1:]
+
+    dates = (
+        pd.DatetimeIndex([dates])
+        if isinstance(dates, pd.Timestamp)
+        else (
+            pd.DatetimeIndex(dates)
+            if isinstance(dates, (pd.Series, pd.DatetimeIndex))
+            else dates.normalize()
         )
-    return dates.normalize()
+    )
+
+    return call._func(*(call._args[0], dates) if method else (dates,), *other_args)
 
 
-def dt2date(dtarr: NDArray[datetime64]) -> NDArray[int32]:
+@decorator
+def to_dt64(
+    call: DatetimeScalarOrArray, freq: str = "D", to_int64: bool = False
+) -> Any:
+    """
+    Decorator to convert date or date array to datetime64 date or array of
+    specified frequency (defaults to day, 'D'). If to_int64 flag is passed,
+    it will ignore frequency argument and return int64 time since the epoch in
+    nanoseconds.
+
+    Arguments
+    ---------
+    call : calling function
+
+    freq : optional string frequency for datetime64 conversion, defaults to
+        'D'. Must be a valid frequency (e.g. 'D', 's', 'us', 'ns')
+
+    to_int64: boolean flag to instead convert output to int64 nanoseconds
+        since the epoch.
+
+    Returns
+    -------
+    Wrapped func output from converted datetime
+    """
+    args = call._args
+
+    if len(args) > 1 and hasattr(args[0], "__class__"):
+        dt = args[1]
+        other_args = args[2:]
+        call_with_method = True
+    else:
+        dt = args[0]
+        other_args = args[1:]
+        call_with_method = False
+
+    if pd.api.types.is_scalar(val=dt):
+        dt = pd.Timestamp(dt).normalize()
+    elif isinstance(dt, pd.Series):
+        dt: DatetimeIndex = pd.DatetimeIndex(data=dt).normalize()
+    else:
+        dt = pd.to_datetime(dt).normalize()
+
+    dt_converted = dt.to_numpy().astype("int64" if to_int64 else f"datetime64[{freq}]")
+    if call_with_method:
+        return call(args[0], dt_converted, *other_args, **call._kwargs)
+    else:
+        return call(dt_converted, *other_args, **call._kwargs)
+
+
+@to_dt64(freq="ns")
+def dt64_to_date(dtarr: NDArray[datetime64]) -> NDArray[int32]:
     """
     Adapted from RBF06:
     https://stackoverflow.com/questions/13648774/get-year-month-or-day-from-numpy-datetime64#26895491
@@ -155,7 +237,6 @@ def dt2date(dtarr: NDArray[datetime64]) -> NDArray[int32]:
     uint32 array (..., 4)
         calendar array with last axis representing year, month, day
     """
-
     out: NDArray[Any] = np.empty(shape=dtarr.shape + (4,), dtype="u4")
 
     Y, M, D = [dtarr.astype(dtype=f"M8[{x}]") for x in "YMD"]
@@ -167,6 +248,7 @@ def dt2date(dtarr: NDArray[datetime64]) -> NDArray[int32]:
     return out
 
 
+@to_dt64
 def dt64_to_dow(dtarr: NDArray[datetime64]) -> NDArray[datetime64 | int64]:
     """
     Adapted from jwdink on stackoverflow:
@@ -185,8 +267,8 @@ def dt64_to_dow(dtarr: NDArray[datetime64]) -> NDArray[datetime64 | int64]:
         array with axis representing day of week
     """
     out: NDArray[uint8] = np.empty(shape=dtarr.shape + (2,), dtype=uint8)
-    out[..., 0] = dtarr.astype(dtype="datetime64[D]")
-    out[..., 1] = (dtarr.astype(dtype="datetime64[D]").view(dtype="int64") - 4) % 7
+    out[..., 0] = dtarr
+    out[..., 1] = (dtarr.view(dtype="int64") - 4) % 7
     return out
 
 
@@ -271,7 +353,7 @@ class YearMonthDay:
         A POSIX-day timestamp as an integer (whole days since the Unix Epoch).
 
         """
-        return pdtimestamp_to_posix_day(timestamp=self.to_pdtimestamp())
+        return ts_to_posix_day(timestamp=self.to_pdtimestamp())
 
     def to_pdtimestamp(self) -> Timestamp:
         """
@@ -293,7 +375,6 @@ class YearMonthDay:
         A Python datetime.date object.
 
         """
-
         return datetime.date(year=self.year, month=self.month, day=self.day)
 
     @property
@@ -338,8 +419,8 @@ def to_timestamp(date_input: FedStampConvertibleTypes) -> Timestamp | None:
 
     """
     raise TypeError(
-        f"""Unsupported date format. You provided type: {type(date_input)}.
-        Supported types are FedStampConvertibleTypes"""
+        f"Unsupported date format. You provided type: {type(date_input)}."
+        "Supported types are FedStampConvertibleTypes"
     )
 
 
@@ -361,6 +442,10 @@ def _posix_to_timestamp(date_input: int | int64 | float) -> Timestamp:
     period. We follow the pandas convention of terminating at year 2200 (84000
     posix-days is actually 2199-12-26, but it's round).
     """
+    if (
+        isinstance(date_input, np.int64) or date_input > 7_258_032_000
+    ):  # year 2200 in epoch seconds
+        return _normalize_timestamp(pd.to_datetime(date_input, unit="ns"))
     if date_input < 84000:
         return _normalize_timestamp(pd.to_datetime(date_input, unit="D"))
     return _normalize_timestamp(pd.to_datetime(date_input, unit="s"))
@@ -389,11 +474,11 @@ def _str_to_timestamp(date_input: str) -> Timestamp:
             except ValueError:
                 continue
         raise ValueError(
-            f"""Date string '{date_input}' is not in a recognized format. All
-            reasonable attempts to parse it failed. Are you trying to use an
-            alien date format? If you communicate in timelessness like those
-            aliens in Arrival we're at an impasse. Please use an ISO 8601
-            format"""
+            f"Date string '{date_input}' is not in a recognized format. All "
+            "reasonable attempts to parse it failed. Are you trying to use an "
+            "alien date format? If you communicate in timelessness like those "
+            "aliens in Arrival we're at an impasse. Please use an ISO 8601 "
+            "format."
         ) from e
 
 
@@ -417,19 +502,19 @@ def _yearmonthday_to_timestamp(date_input: YearMonthDay) -> Timestamp:
 def _timetuple_to_timestamp(date_input: tuple) -> Timestamp:
     if len(date_input) != 3:
         raise ValueError(
-            """Timetuple input requires a tuple with four-digit year, month,
-            day as integers or integer-convertible strings."""
+            "Timetuple input requires a tuple with four-digit year, month, "
+            "day as integers or integer-convertible strings."
         )
     try:
         year, month, day = (int(item) for item in date_input)
     except ValueError as e:
         raise ValueError(
-            """Year, month, and day must be integers or strings that can be
-            converted to integers."""
+            "Year, month, and day must be integers or strings that can be "
+            "converted to integers."
         ) from e
 
     if not 1970 <= year <= 2200:
-        raise ValueError("Year must be a four-digit number, and not before 1970.")
+        raise ValueError("Year must be a four-digit number between 1970 and 2199")
 
     return _normalize_timestamp(
         YearMonthDay(year=year, month=month, day=day).to_pdtimestamp()
@@ -442,22 +527,21 @@ def _check_year(dates: Timestamp | DatetimeIndex) -> Timestamp | DatetimeIndex:
             return dates
         else:
             raise ValueError(
-                f"Input dates must be in range 1970-1-1 and 2199-12-31, you provided a date from year {dates.year}."
+                "Input dates must be in range 1970-1-1 and 2199-12-31, you "
+                f"provided a date from year {dates.year}."
             )
 
     if dates.year.min() > 1969 and dates.year.max() < 2200:
         return dates
     else:
         raise ValueError(
-            f"""Input dates must be in range 1970-1-1 and 2199-12-31.
-            You provided dates in range {dates.year.min()}-{dates.year.max()}.
-            """
+            "Input dates must be in range 1970-1-1 and 2199-12-31. "
+            f"You provided dates in range {dates.year.min()}-{dates.year.max()}."
         )
 
 
-def check_timestamp(
-    func,
-) -> Callable[..., Any] | Any | Timestamp | None:
+@decorator
+def check_timestamp(call) -> Timestamp | None:
     """
     Since _normalize_timestamp is designed to normalize Timestamps, to avoid
     repeating ourselves with conversions to Timestamps in most of our
@@ -472,34 +556,28 @@ def check_timestamp(
     Returns
     -------
     A wrapper around func that converts non-pd.Timestamp input to Timestamps.
-
     """
+    arg = call._args[0]
 
-    @wraps(wrapped=func)
-    def wrapper(arg) -> Timestamp | Any | None:
-        """Our pd.Timestamp handling wrapper."""
-        if isinstance(arg, pd.Timestamp):
-            return func(_check_year(dates=arg))
+    if isinstance(arg, pd.Timestamp):
+        return call(_check_year(dates=arg))
 
-        if arg is None:
-            raise ValueError(
-                f"""provided argument, {arg} is None; we're not mind readers
-                here. Please provide a pd.Timestamp for _normalize_timestamp.
-                """
-            )
-        try:
-            ts = pd.Timestamp(arg)
-            return func(_check_year(dates=ts))
-        except TypeError as e:
-            raise TypeError(
-                f"""input {arg} could not be converted to a pd.Timestamp.
-                    Our _normalize_timestamp function needs pandas Timestamps
-                    or a pd.Timestamp convertible date-like object (e.g. Python
-                    date, timetuple, string-date, POSIX)
-                    """
-            ) from e
+    if arg is None:
+        raise ValueError(
+            f"provided argument, {arg} is None; we're not mind readers "
+            "here. Please provide a pd.Timestamp for _normalize_timestamp."
+        )
 
-    return wrapper
+    try:
+        ts = pd.Timestamp(arg)
+        return call(_check_year(dates=ts))
+    except TypeError as e:
+        raise TypeError(
+            f"input {arg} could not be converted to a pd.Timestamp. "
+            "Our _normalize_timestamp function needs pandas Timestamps "
+            "or a pd.Timestamp convertible date-like object (e.g. Python "
+            "date, timetuple, string-date, POSIX)"
+        ) from e
 
 
 @check_timestamp
@@ -523,9 +601,8 @@ def _normalize_timestamp(timestamp: Timestamp | None = None) -> Timestamp:
     return timestamp.normalize()
 
 
-def wrap_tuple(
-    func,
-) -> Callable[..., Any] | Any | tuple[Any, ...] | DatetimeIndex | None:
+@decorator
+def wrap_tuple(call) -> tuple[Any] | Any | None:
     """
     To avoid repeating ourselves with date converters that handle two
     arguments for FedIndex, we instead wrap the singledispatch
@@ -547,15 +624,14 @@ def wrap_tuple(
     A wrapper around func that converts multi-argument dates to a tuple.
 
     """
+    args = call._args
 
-    @wraps(wrapped=func)
-    def wrapper(*args) -> Callable | tuple[Any] | Any | None:
-        """Our to-tuple handling wrapper."""
-        if count := len(args) in {1, 2}:
-            return func(args[0]) if count == 1 else func((args[0], args[1]))
-        raise ValueError(f"We need 1 or 2 arguments... not {count}")
-
-    return wrapper
+    if len(args) == 1:
+        return call(args[0])
+    elif len(args) == 2:
+        return call((args[0], args[1]))
+    else:
+        raise ValueError(f"Expected 1 or 2 arguments, got {len(args)}")
 
 
 @wrap_tuple
@@ -588,7 +664,9 @@ def to_datetimeindex(*input_dates: FedIndexConvertibleTypes) -> DatetimeIndex | 
     """
 
     raise TypeError(
-        "You provided unsupported types. Supported types are FedIndexConvertibleTypes"
+        "You provided unsupported types. "
+        "Supported types are FedIndexConvertibleTypes: \n"
+        f"{FedIndexConvertibleTypes}"
     )
 
 
@@ -689,3 +767,19 @@ def _normalize_datetimeindex(datetimeindex: DatetimeIndex) -> DatetimeIndex:
     if datetimeindex.tz:
         return datetimeindex.tz_convert(tz="UTC").normalize()
     return datetimeindex.normalize()
+
+
+__all__: list[str] = [
+    "YearMonthDay",
+    "check_timestamp",
+    "dt64_to_date",
+    "dt64_to_dow",
+    "ensure_datetimeindex",
+    "get_today",
+    "iso_to_ts",
+    "to_datetimeindex",
+    "to_dt64",
+    "to_timestamp",
+    "ts_to_posix_day",
+    "wrap_tuple",
+]
